requirements.txt
pytest==7.4.0
selenium==4.1.5
pytest-html==3.1.1
pytest-xdist==2.5.0
allure-pytest==2.9.45
webdriver-manager==4.0.0